Fix filter.format.github_flavored_markdown not getting enabled on install
Commit 1cf4a8 tried to fix this issue once by moving the enabling from
hook_install() to hook_modules_installed(). This turned out to be only a
partial solution to the problem of
filter.format.github_flavored_markdown not getting enabled when using
config_installer to install a site. FilterForma::load() cannot be used
when installing with config_installer. Instead a direct config edit has
to be used. This can be done in hook_install(), without misusing
hook_modules_installed().