Remove unsetValue() from the api ref validation
The problem with unsetValue() is that any code calling getValue() will
get a null. There are various parts in Drupal core (e.g. FileWidget)
that thinks that getValue() will always return an array, and puts the
returned value into a foreach() without a null check. This will cause
ugly warnings in the logs.

What this code does instead is that it replaces the unsetValue() calls
to setValue('field_source_file', []), so FileWidget will get an empty
array.